==> code/1_Data_Clean.ipynb <==
# Original Data Cleaning

import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 500)

df=pd.read_csv("../data/A1c_Data_Requesta1c.csv")

## What does our data look like? 

When cleaning the data, we need to consider the following:
1. Data normalization: a process where data in a database is organized in a way that the user can utilize it to conduct analysis and answer questions. This process entails eliminating redundancy and logically group data together. This data was not normalized, there are two dataframes grouped together in one in a way that doesn't make sense (one responsible for a1c and when bloodwork was done, and the other holding information only on office visits). There are also many duplicates in the data. 
2. We have 273,485 rows of data and 8 columns. This data represents 1,099 patients. 
3. Each patient has multiple types of insurances listed, and we are unsure which is the most updated one or what many of them mean. For this reason, I don't believe we may have to eliminate this column for further analysis. 

df.head()

How many patients do we have in this dataset? 

#Check number of patients by ID
df["PatientId"].value_counts()

#change column names so we can more easily work with columns
df.rename(columns={"PatientId": "patient_id","Age": "age", "VisitDate": "visit_date", "SUMMARY":"summary","OBSDATE":"obs_date","Date of A1c":"a1c","Insurance":"insurance"},inplace=True)

#check rows and columns 
df.shape

#check to see what types of insurance exist in the data
df["insurance"].value_counts()

#check how many null rows we have 
df.isnull().sum()

#drop rows where a1c is missing since it is only 32 out of 273485
df.dropna(subset=["a1c"],inplace=True)

df.reset_index(inplace=True)

#check to make sure the null values were deleted
df.isnull().sum()

#drop insurance and birthday 
df_no_ins= df.drop(columns=["insurance","Birthdate"])

#check datatypes: are the columns meant to be numerical, numerical? Are the dates in the dataframe datetime datatypes? 
df_no_ins.dtypes

#convert visit date and obs date to datetime, and take out time 
df_no_ins['visit_date'] = pd.to_datetime(df_no_ins['visit_date'])

df_no_ins['visit_date']= df_no_ins['visit_date'].dt.date

df_no_ins['visit_date'] = pd.to_datetime(df_no_ins['visit_date'])

df_no_ins['obs_date'] = pd.to_datetime(df_no_ins['obs_date'])

df_no_ins['obs_date'] = df_no_ins['obs_date'].dt.date

df_no_ins['obs_date'] = pd.to_datetime(df_no_ins['obs_date'])

df_no_ins.dtypes

df_no_ins["a1c"].sort_values(ascending=True)

#.089 is too low
df_no_ins["a1c"].min()

df_no_ins["a1c"].max()

#mean a1c value is 8.9
df_no_ins["a1c"].describe()

#replace .089 with 8.9 
df_no_ins["a1c"].where(df_no_ins["a1c"]!= 0.08900000000000001, 8.9,inplace=True)

df_no_ins.drop(columns=["index"],inplace=True)

df_no_ins["a1c"].min()

#creating a mask so we can drop 
mask= df_no_ins.loc[df_no_ins["a1c"]== 1.7]

mask

mask.index

df_no_ins.shape

df_no_ins.drop(mask.index,axis=0,inplace=True)

df_no_ins["a1c"].min()

df_no_ins.shape

#time to drop duplicates!
df_no_ins.drop_duplicates(inplace=True)

df_no_ins.reset_index(drop=True,inplace=True)

df_no_ins.shape
# dropped 210,958 

Add column to dummy a1c level so we can use both classification and regression models 

df_no_ins["a1c_level"]=""

for i in range(df_no_ins.shape[0]):
    if df_no_ins["a1c"][i] <=8:
        df_no_ins["a1c_level"][i]= 1
    else: df_no_ins["a1c_level"][i]=0
        #1 is controlled
        #0 is uncontrolled 

df_no_ins.head()

#save to csv so we can use it for splitting and EDA
df_no_ins.to_csv("../data/cleaned_data.csv")
