==> code/2_Data_Clean_and_Split.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 500)

new_data= pd.read_csv("../data/cleaned_data.csv")

new_data.head()

new_data.drop(columns=["Unnamed: 0"],inplace=True)

new_data.dtypes
#check the datatypes

#change datatypes to correct type 
new_data['visit_date'] = pd.to_datetime(new_data['visit_date'])

new_data['obs_date'] = pd.to_datetime(new_data['obs_date'])

new_data.dtypes

#count the number of times they got their bloodwork done
new_data["num_obs"]=""

for i in range(new_data.shape[0]):
    try:
        mas=new_data.loc[new_data['patient_id'] == new_data['patient_id'][i]]
        new_data["num_obs"][i]= mas['obs_date'].nunique()
    except: 
        continue 

#count the number of visits 
new_data["num_visits"]=""

for i in range(new_data.shape[0]):
    try:
        mas=new_data.loc[new_data['patient_id'] == new_data['patient_id'][i]]
        new_data["num_visits"][i]= mas["visit_date"].nunique()
    except: 
        continue

#split the data into two dataframes since dates don't line up
summary_and_visit_df = new_data[['patient_id', 'age', 'visit_date','summary',"num_visits"]].copy()
date_a1c_df= new_data[['patient_id', 'age', 'obs_date','a1c','a1c_level',"num_obs"]].copy()

date_a1c_df.shape

summary_and_visit_df.shape

date_a1c_df.drop_duplicates(inplace=True)

date_a1c_df.shape

summary_and_visit_df.drop_duplicates(inplace=True)

summary_and_visit_df.shape

summary_and_visit_df.to_csv("../data/summary_and_visit_df.csv")

date_a1c_df.to_csv("../data/date_a1c_df.csv")
