==> code/3_Feature_Engineering.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 500)

summary_and_visit_df= pd.read_csv("../data/summary_and_visit_df")

sorted_sum= summary_and_visit_df.sort_values(by=["patient_id", "visit_date"])

sorted_sum.drop(columns=["Unnamed: 0"],inplace=True)

sorted_sum.reset_index(inplace=True,drop=True)

sorted_sum.dtypes

sorted_sum["visit_date"]=pd.to_datetime(sorted_sum["visit_date"])

Get the time in between visits

sorted_sum["time_between_visits"]=""

for i in range(sorted_sum.shape[0]):
    try:
        patient_sum= sorted_sum.loc[sorted_sum['patient_id'] == sorted_sum['patient_id'][i]]
        sorted_sum["time_between_visits"][i]= patient_sum["visit_date"][i+1] - patient_sum["visit_date"][i]
    except: 
        i= i+1

sorted_sum.dropna(subset=["time_between_visits"],inplace=True)

Turn datetime object to a float so we can use it for modelling 

def split_dates(dates_obj):
    if dates_obj:
        to_string=str(dates_obj)
        date_num= to_string.split()        
        return float(date_num[0])
    else: 
        return 

sorted_sum["time_between_visits"]= sorted_sum["time_between_visits"].map(split_dates)

Get the average time between visits for each patient

values_time= sorted_sum.groupby(by='patient_id',as_index=False)['time_between_visits'].mean()

Create a column to add average time between visits for each patient

sorted_sum["avg_time_between_visits"]=""

for x in range(sorted_sum.shape[0]):
    for i in range(len(values_time)):
        try:
            if sorted_sum["patient_id"][x]== values_time["patient_id"][i]:
                #index of where patient id is that value 
                sorted_sum["avg_time_between_visits"][x]= values_time["time_between_visits"][i]
        except:
            continue

sorted_sum.to_csv("../data/sorted_summary_with_avg_visits.csv")
