==> code/4_Feature_Engineering.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 500)

sorted_sum=pd.read_csv("../data/sorted_summary_with_avg_visits.csv")

date_a1c_df=pd.read_csv("../data/date_a1c_df.csv")

sorted_sum.head()

sorted_sum.drop(columns=["Unnamed: 0"],inplace=True)

sorted_sum.shape

date_a1c_df.head()

date_a1c_df.drop(columns=["Unnamed: 0"],inplace=True)

avg_a1c= date_a1c_df.groupby('patient_id',as_index=False)['a1c'].mean()

avg_a1c

I am adding average a1c for each patient as a feature. 

sorted_sum["avg_a1c"]=""

for x in range(sorted_sum.shape[0]):
    for i in range(len(avg_a1c)):
        try:
            if sorted_sum["patient_id"][x]== avg_a1c["patient_id"][i]:
                #index of where patient id is that value 
                sorted_sum["avg_a1c"][x]= avg_a1c["a1c"][i]
        except:
            continue 

#save this to a csv
sorted_sum.to_csv("../data/sorted_sum_data.csv")
