==> code/5_Feature_Engineering.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
pd.set_option('display.max_columns', 100)
pd.set_option('display.max_rows', 500)
%matplotlib inline

df= pd.read_csv("../data/sorted_sum_data.csv")

data_a1c_df=pd.read_csv("../data/date_a1c_df.csv")

df.head()

df.drop(columns=["Unnamed: 0"],inplace=True)

#create a new dataframe so that we have one row per patient with average a1c,average time between visits, age, and number of visits
new_df = df[['patient_id', 'age',"num_visits","avg_time_between_visits","avg_a1c"]].copy()

new_df.head()

new_df.drop_duplicates(inplace=True)

new_df.reset_index(inplace=True,drop=True)

new_df.head()

Add column to indicate whether the patient's average a1c is classified as our positive class (1,controlled) or negative class (0,uncontrolled)

new_df["avg_control_level"]=""

for i in range(new_df.shape[0]):
    if new_df["avg_a1c"][i] <=8:
        new_df["avg_control_level"][i]= 1
    else: new_df["avg_control_level"][i]=0
        #1 is controlled

data_a1c_df.head()

# add it number of observations per patient 

new_df["num_obs"]=""

for x in range(data_a1c_df.shape[0]):
    for i in range(new_df.shape[0]):
            if data_a1c_df["patient_id"][x]== new_df["patient_id"][i]:
                #index of where patient id is that value 
                new_df["num_obs"][i]= data_a1c_df["num_obs"][x]

new_df.isnull().sum()

new_df.head()

new_df.to_csv("../data/new_df_for_modelling.csv")
