==> code/7_Modeling.ipynb <==
import pandas as pd
import numpy as np
from sklearn.linear_model import LinearRegression, LassoCV, RidgeCV,Ridge, Lasso,LogisticRegression
from sklearn.preprocessing import LabelBinarizer
from sklearn.preprocessing import PolynomialFeatures, StandardScaler
from sklearn.model_selection import train_test_split, cross_val_score, cross_val_predict
from sklearn.metrics import r2_score
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import GridSearchCV
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import roc_curve, precision_recall_curve, auc, make_scorer, recall_score, accuracy_score, precision_score, confusion_matrix
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split, GridSearchCV, StratifiedKFold

new_df= pd.read_csv("../data/new_df_for_modelling.csv")

new_df.head()

new_df.drop(columns=["Unnamed: 0"],inplace=True)

new_df.dtypes

convert = lambda x: round(x)
#round a1c levels for regression model

new_df.head()

new_df["rounded_a1c"]= new_df["avg_a1c"].map(convert)

new_df["avg_control_level"].value_counts(normalize=True)
#baseline score 

new_df.isnull().sum()

new_df.dropna(inplace=True)

new_df.isnull().sum()

I decided to try both regression and classification modeling. Regression would be used to predict the patient's average a1c score. Classification would be used to predict whether or not the patient on average was controlled or uncontrolled.  

## Linear Regression

How well does the model predict average a1c if we just use average time between visits and number of visits? 

features=['avg_time_between_visits',"num_visits"]
X = new_df[features] # X = what we use to predict y
y = new_df['avg_a1c'] # y = what we want to predict

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

lr = LinearRegression()
lr.fit(X_train, y_train)

print(f'Score on the training set: {lr.score(X_train, y_train)}')
print(f'Score on the test set: {lr.score(X_test, y_test)}')

What happens when we add age and number of observations to our features list? 

features=['avg_time_between_visits',"age","num_visits","num_obs"]
X = new_df[features] # X = what we use to predict y
y = new_df['avg_a1c'] # y = what we want to predict

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

lr = LinearRegression()
lr.fit(X_train, y_train)

print(f'Score on the training set: {lr.score(X_train, y_train)}')
print(f'Score on the test set: {lr.score(X_test, y_test)}')

Our Linear Regression model isn't doing a great job at predicting average a1c, but let's try this with rounding our a1c levels. 

features=['avg_time_between_visits',"age","num_visits","num_obs"]
X = new_df[features] # X = what we use to predict y
y = new_df['rounded_a1c'] # y = what we want to predict

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)

lr = LinearRegression()
lr.fit(X_train, y_train)

print(f'Score on the training set: {lr.score(X_train, y_train)}')
print(f'Score on the test set: {lr.score(X_test, y_test)}')

Our score does not improve after using rounded a1c levels. It is also worth noting that with this type of data, it is unlikely that all MLR assumptions are being met. 

## Logistic Regression

Let's see if we can better predict average control level (whether or not on average, a patient is controlled or uncontrolled)

features=['avg_time_between_visits',"age","num_visits","num_obs"]
X = new_df[features]
y = new_df['avg_control_level']

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=39)

logreg = LogisticRegression()

logreg.fit(X_train, y_train)

print(f'Score on the training set: {logreg.score(X_train, y_train)}')
print(f'Score on the test set: {logreg.score(X_test, y_test)}')

print(f'Logistic Regression Intercept: {logreg.intercept_}')
print(f'Logistic Regression Coefficient: {logreg.coef_}')

np.exp(logreg.coef_)

The score we get here isn't significantly better than our baseline score of .60 in the negative class, but we can make a few (relatively unhelpful) inferences: 

- As average time between visits increases by 1 day, someone is about 1.002 times as likely to be controlled.
- As age increases by 1 year, someone is about 1.002 times as likely to be controlled.
- As number of visits increases by 1, someone is about 1.04 times as likely to be controlled.
- As person has more observations done, they are less likely to be in the positive class/ controlled (.36 times as likely to be negative class) 



## Decision Tree Classifier

Here I used gridsearch to find the best estimator and hyperparameters. 

grid = GridSearchCV(estimator = DecisionTreeClassifier(),
                    param_grid = {'max_depth': [3, 5, 7, 10],
                                  'min_samples_split': [5, 10, 15, 20],
                                  'min_samples_leaf': [2, 3, 4, 5, 6, 7]},
                    cv = 5,
                    verbose = 1)

grid.fit(X_train, y_train)

grid.best_estimator_

grid.best_score_

dt = grid.best_estimator_

# Fit model.
dt.fit(X_train, y_train)

# Evaluate model.
print(f'Score on training set: {dt.score(X_train, y_train)}')
print(f'Score on testing set: {dt.score(X_test, y_test)}')

preds = dt.predict(X_test)

tn, fp, fn, tp = confusion_matrix(y_test,
                                  preds).ravel()

print(confusion_matrix(y_test,
                       preds))

# Calculate sensitivity.

sens = tp / (tp + fn)

print(f'Sensitivity: {round(sens, 4)}')
# tell it to use specificity insteead of accuracy 

# Calculate specificity.

spec = tn / (tn + fp)

print(f'Specificity: {round(spec, 4)}')

In a healthcare setting, usually you want to optimize for sensitivity. Here, our specificity score (optimizing for true negatives) is much higher than our sensitivity score (optimizing for true positives). We would rather tell a patient who is controlled that they are not as opposed to telling a patient they are controlled when they are not. 


# Random Forest Classifier

With Random Forest Classifier, we will attempt to optimize for sensitivity. 

features=['avg_time_between_visits',"age","num_visits","num_obs"]
X = new_df[features]
y= new_df["avg_control_level"]

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y)

#The following code is taken from Kevin Arvai's post on how to optimize for sensitivity: https://towardsdatascience.com/fine-tuning-a-classifier-in-scikit-learn-66e048c21e65

clf = RandomForestClassifier(n_jobs=-1)

param_grid = {
    'min_samples_split': [3, 5, 10], 
    'n_estimators' : [100, 300],
    'max_depth': [3, 5, 15, 25],
    'max_features': [0.1, .25, 0.5, 1.0]
}

scorers = {
    'precision_score': make_scorer(precision_score),
    'recall_score': make_scorer(recall_score),
    'accuracy_score': make_scorer(accuracy_score)
}

#The following code is taken from Kevin Arvai's post on how to optimize for sensitivity: https://towardsdatascience.com/fine-tuning-a-classifier-in-scikit-learn-66e048c21e65 
def grid_search_wrapper(refit_score='precision_score'):
    """
    fits a GridSearchCV classifier using refit_score for optimization
    prints classifier performance metrics
    """
    skf = StratifiedKFold(n_splits=10)
    grid_search = GridSearchCV(clf, param_grid, scoring=scorers, refit=refit_score,
                           cv=skf, return_train_score=True, n_jobs=-1)
    grid_search.fit(X_train.values, y_train.values)

    # make the predictions
    y_pred = grid_search.predict(X_test.values)

    print('Best params for {}'.format(refit_score))
    print(grid_search.best_params_)

    # confusion matrix on the test data.
    print('\nConfusion matrix of Random Forest optimized for {} on the test data:'.format(refit_score))
    print(pd.DataFrame(confusion_matrix(y_test, y_pred),
                 columns=['pred_neg', 'pred_pos'], index=['neg', 'pos']))
    return grid_search

#The following code is taken from Kevin Arvai's post on how to optimize for sensitivity: https://towardsdatascience.com/fine-tuning-a-classifier-in-scikit-learn-66e048c21e65
grid_search_clf = grid_search_wrapper(refit_score='precision_score')

#The following code is taken from Kevin Arvai's post on how to optimize for sensitivity: https://towardsdatascience.com/fine-tuning-a-classifier-in-scikit-learn-66e048c21e65 
results = pd.DataFrame(grid_search_clf.cv_results_)
results = results.sort_values(by='mean_test_precision_score', ascending=False)
results[['mean_test_precision_score', 'mean_test_recall_score', 'mean_test_accuracy_score', 'param_max_depth', 'param_max_features', 'param_min_samples_split', 'param_n_estimators']].round(3).head()

#The following code is taken from Kevin Arvai's post on how to optimize for sensitivity: https://towardsdatascience.com/fine-tuning-a-classifier-in-scikit-learn-66e048c21e65
grid_search_clf = grid_search_wrapper(refit_score='recall_score')

#The following code is taken from Kevin Arvai's post on how to optimize for sensitivity: https://towardsdatascience.com/fine-tuning-a-classifier-in-scikit-learn-66e048c21e65
results = pd.DataFrame(grid_search_clf.cv_results_)
results = results.sort_values(by='mean_test_precision_score', ascending=False)

#The following code is taken from Kevin Arvai's post on how to optimize for sensitivity: https://towardsdatascience.com/fine-tuning-a-classifier-in-scikit-learn-66e048c21e65
results[['mean_test_precision_score', 'mean_test_recall_score', 'mean_test_accuracy_score', 'param_max_depth', 'param_max_features', 'param_min_samples_split', 'param_n_estimators']].round(3).head()

As we attempt to optimize for recall, we still get a relatively low score
